Force screen to stay on while logging
This overrides any auto lock setting so the screen won't go to sleep. This is useful since when the screen goes to sleep I've observed the advertised name and TX power disappear. Note that the proximity sensor is still functional and can turn the screen off, but in that case we can still advertise (since the app doesn't go into the background).