Update compass for 15 degree resolution
Only logs changes very 15 degrees (the default 1 degree is likely way more than we need). Also added a note about location accuracy for navigation...this could be used for better estimates, as it uses other sensors on the phone, but Apple only recommends using it if the device is plugged in to power.