Detector: add time threshold
This time threshold is used to decide to show a device in the table or not. For a threshold T, if the device has not been seen in T seconds, it will no longer be displayed and its data will be reset (so when it returns it's like a fresh look).