Ultrasonic: add Bluetooth control to start/stop
Device A acts like the master, and device B the slave. When the user starts ultrasonic processing on A, a Bluteooth advertisement is sent out so that B also knows to start (and same goes for stopping). The Bluetooth scan on B is designed to only return the first command from A.

Also fix disable/enable of UI elements based on if processing is running or not.