Ultrasonic: do one measurement per run
Changed measurement procedure so it'll be easier to test ranging on the phones. The master initiates a measurement over Bluetooth and transmits first. The slave transmits after waiting 300 ms. Both devices start recording as soon as they can, and record for 682 ms. The transmit waveforms are stil 200 ms long, so there should be plenty of margin.

Other small changes:
- Added counter so we can see how many measurements have been done
- Disable run button while running
- Change run button text to "Waiting..." on slave device