Ultrasonic: add measurement exchange via BLE
Uses the BLE advertised name as a kludgy way to transfer measurements between the master/slave. Currently just transfering the time the waveform was sent from each device (not actually doing any ranging processing). Updated the UI to show the times as well as range, for whenever we can actually compute it.