Ultrasonic: update timing for 19200 sample buffer
Appears that the previous buffer length didn't work, as only 19200 samples (0.4 seconds) are returned. Trying to keep the data to be processed to one buffer to simplify things - so reduced waveform to 100 ms and changed the delay to transmit for both master and slave nodes. Received data looks reasonable in MATLAB, although I've only looked at the time series and crosscorrelation, and not actually done the full two-way ranging processing.

Also changed the order running the TX/RX workers, since it seems like there's a decent delay before the waveform is actually transmitted (should be ~50 ms, but was closer to 100 ms - and after the change, it's closer to 50 ms again).