Ultrasonic: initial cut at matched filter
Current just processes for its own transmit signal and estimates the loopback delay. Implemented an xcorr that behaves more like MATLAB. Results on device are similar to those in MATLAB, although it appears that the loopback delays for the "master" device are no longer as consistent as they were in the past. This might be fine for the relative timing measurements, but also possible there is something messed up with the xcorr or peak detection on device.