Ultrasonic: make separate master/slave waveforms
Each device makes both waveforms so they can do the measurements. Currently only loopback delay is measured and transferred over BLE. It seems to work most of the time, but sometimes the counter on the slave gets ahead by one and it doesn't show it's own measurement - not quite sure why. There were also crashes when trying to run too soon after the last run was done, that's mostly fixed, although I haven't tested it enough to be super confident in it.