Ultrasonic: initial range calculation on device!
Now correlating with the remote waveform and doing the full two way ranging calculation. Initial tests at 1 and 2 feet show it is working well, but usually only for the first run after Bluetooth is cycled (via the settings menu, not the controls that pull up from the bottom). Not sure why that is, but if I cycle BLE on both devices, then start up the app and make a measurement, it's usually very accurate.