Ultrasonic: add basic SNR estimator
Estimates noise before the master transmits, and signal from the peak output of the matched filter. Results agree with MATLAB, though I noticed that if one device does not transmit (i.e. only master making measurements on itself), the correlation peak is from the tail end of the device itself transmitting. So the estimate is not that accurate and we may want to apply further windowing for peak selection (based on the expected location).