Ultrasonic: fix BLE control flow
The BLE control and data sharing seemed pretty unreliable - namely, the master would stop being able to send the uStart signal, and both nodes were sometimes receiving multiple copies of the same measurement (or slightly different ones). I think the primary issue was putting the full double precision measurement (with decimal place) in the name. After rounding it to the nearest nanosecond things seem much more stable. Also added protection against starting a run while we're already running (for the slave), and overwriting a measurement if a duplicate one is returned by the scan.

This update includes running states for the advertiser/scanner and makes sure to stop them before changing the name (depends on which node we are). Things are also left in a good state when we go to other tabs in the app (i.e. not stuck in ultrasonic mode).