Ultrasonic: improved results in multipath
Changed the detector to use an adaptive threshold based on total power level, followed by a peak search over some small window of samples (to find the true peak, or first peak in the case of multipath). The parameters were set based on a small amount of indoor and outdoor data.